move update_platform to convert/add_ancillary move update_platform to under convert/add_ancillary, need to fix function and broken dependencies
class-redesign' into echodata-overhaul
