+++ echopype/convert/convert.py
@@ -3,11 +3,7 @@ UI class for converting raw data from different echosounders to netcdf or zarr.
 """
 import os
 import warnings
-import xarray as xr
-import numpy as np
-import zarr
 from pathlib import Path
-from collections.abc import MutableMapping
 from datetime import datetime as dt
 import fsspec
 from fsspec.implementations.local import LocalFileSystem
@@ -383,123 +379,6 @@ class Convert:
             for f in indiv_files:
                 combine_fcn.remove_indiv_files(f)
 
-    def update_platform(self, files=None, extra_platform_data=None):
-        """
-        Parameters
-        ----------
-        files : str / list
-            path of converted .nc/.zarr files
-        extra_platform_data : xarray dataset
-            dataset containing platform information along a 'time' dimension
-        """
-        # self.extra_platform data passed into to_netcdf or from another function
-        if extra_platform_data is None:
-            return
-        files = self.output_file if files is None else files
-        if not isinstance(files, list):
-            files = [files]
-        engine = io.get_file_format(files[0])
-
-        # saildrone specific hack
-        if "trajectory" in extra_platform_data:
-            extra_platform_data = extra_platform_data.isel(trajectory=0).drop("trajectory")
-            extra_platform_data = extra_platform_data.swap_dims({'obs': 'time'})
-
-        # Try to find the time dimension in the extra_platform_data
-        possible_time_keys = ['time', 'ping_time', 'location_time']
-        time_name = ''
-        for k in possible_time_keys:
-            if k in extra_platform_data:
-                time_name = k
-                break
-        if not time_name:
-            raise ValueError('Time dimension not found')
-
-        for f in files:
-            ds_beam = xr.open_dataset(f, group="Beam", engine=engine)
-            ds_platform = xr.open_dataset(f, group="Platform", engine=engine)
-
-            # only take data during ping times
-            # start_time, end_time = min(ds_beam["ping_time"]), max(ds_beam["ping_time"])
-            start_time, end_time = ds_beam.ping_time.min(), ds_beam.ping_time.max()
-
-            extra_platform_data = extra_platform_data.sel({time_name: slice(start_time, end_time)})
-
-            def mapping_get_multiple(mapping, keys, default=None):
-                for key in keys:
-                    if key in mapping:
-                        return mapping[key]
-                return default
-
-            if self.sonar_model in ['EK80', 'EA640']:
-                ds_platform = ds_platform.reindex({
-                    "mru_time": extra_platform_data[time_name].values,
-                    "location_time": extra_platform_data[time_name].values,
-                })
-                # merge extra platform data
-                num_obs = len(extra_platform_data[time_name])
-                ds_platform = ds_platform.update({
-                    "pitch": ("mru_time", mapping_get_multiple(
-                        extra_platform_data, ["pitch", "PITCH"], np.full(num_obs, np.nan))),
-                    "roll": ("mru_time", mapping_get_multiple(
-                        extra_platform_data, ["roll", "ROLL"], np.full(num_obs, np.nan))),
-                    "heave": ("mru_time", mapping_get_multiple(
-                        extra_platform_data, ["heave", "HEAVE"], np.full(num_obs, np.nan))),
-                    "latitude": ("location_time", mapping_get_multiple(
-                        extra_platform_data, ["lat", "latitude", "LATITUDE"], default=np.full(num_obs, np.nan))),
-                    "longitude": ("location_time", mapping_get_multiple(
-                        extra_platform_data, ["lon", "longitude", "LONGITUDE"], default=np.full(num_obs, np.nan))),
-                    "water_level": ("location_time", mapping_get_multiple(
-                        extra_platform_data, ["water_level", "WATER_LEVEL"], np.ones(num_obs)))
-                })
-            elif self.sonar_model == 'EK60':
-                ds_platform = ds_platform.reindex({
-                    "ping_time": extra_platform_data[time_name].values,
-                    "location_time": extra_platform_data[time_name].values,
-                })
-                # merge extra platform data
-                num_obs = len(extra_platform_data[time_name])
-                ds_platform = ds_platform.update({
-                    "pitch": ("ping_time", mapping_get_multiple(
-                        extra_platform_data, ["pitch", "PITCH"], np.full(num_obs, np.nan))),
-                    "roll": ("ping_time", mapping_get_multiple(
-                        extra_platform_data, ["roll", "ROLL"], np.full(num_obs, np.nan))),
-                    "heave": ("ping_time", mapping_get_multiple(
-                        extra_platform_data, ["heave", "HEAVE"], np.full(num_obs, np.nan))),
-                    "latitude": ("location_time", mapping_get_multiple(
-                        extra_platform_data, ["lat", "latitude", "LATITUDE"], default=np.full(num_obs, np.nan))),
-                    "longitude": ("location_time", mapping_get_multiple(
-                        extra_platform_data, ["lon", "longitude", "LONGITUDE"], default=np.full(num_obs, np.nan))),
-                    "water_level": ("location_time", mapping_get_multiple(
-                        extra_platform_data, ["water_level", "WATER_LEVEL"], np.ones(num_obs)))
-                })
-
-            # need to close the file in order to remove it
-            # (and need to close the file so to_netcdf can write to it)
-            ds_platform.close()
-            ds_beam.close()
-
-            if engine == "netcdf4":
-                # https://github.com/Unidata/netcdf4-python/issues/65
-                # Copy groups over to temporary file
-                # TODO: Add in documentation: recommended to use Zarr if using add_platform
-                new_dataset_filename = f + ".temp"
-                groups = ['Provenance', 'Environment', 'Beam', 'Sonar', 'Vendor']
-                with xr.open_dataset(f) as ds_top:
-                    ds_top.to_netcdf(new_dataset_filename, mode='w')
-                for group in groups:
-                    with xr.open_dataset(f, group=group) as ds:
-                        ds.to_netcdf(new_dataset_filename, mode='a', group=group)
-                ds_platform.to_netcdf(new_dataset_filename, mode="a", group="Platform")
-                # Replace original file with temporary file
-                os.remove(f)
-                os.rename(new_dataset_filename, f)
-            elif engine == "zarr":
-                # https://github.com/zarr-developers/zarr-python/issues/65
-                old_dataset = zarr.open_group(f, mode="a")
-                del old_dataset["Platform"]
-                ds_platform.to_zarr(f, mode="a", group="Platform")
-
     def _normalize_path(self, out_f, convert_type):
         if convert_type == 'zarr':
             return fsspec.get_mapper(out_f, **self._output_storage_options)
@@ -507,7 +386,7 @@ class Convert:
             return out_f
 
     def _to_file(self, convert_type, save_path=None, data_type='ALL', compress=True, combine=False,
-                 overwrite=False, parallel=False, extra_platform_data=None, storage_options={}, **kwargs):
+                 overwrite=False, parallel=False, storage_options={}, **kwargs):
         """Convert a file or a list of files to netCDF or zarr.
 
         Parameters
@@ -530,8 +409,6 @@ class Convert:
             Defaults to ``False``
         parallel : bool
             whether or not to use parallel processing. (Not yet implemented)
-        extra_platform_data : Dataset
-            The dataset containing the platform information to be added to the output
         storage_options : dict
             Additional keywords to pass to the filesystem class.
         """
@@ -548,7 +425,7 @@ class Convert:
             self._validate_path('.zarr', save_path)
         else:
             raise ValueError('Unknown type to convert file to!')
-        
+
 
         # Get all existing files
         exist_list = []
@@ -582,10 +459,6 @@ class Convert:
         if self.combine:
             self.combine_files(save_path=save_path, remove_indiv=True)
 
-        # Attached platform data
-        if extra_platform_data is not None:
-            self.update_platform(files=self.output_file, extra_platform_data=extra_platform_data)
-
         # If only one output file make it a string instead of a list
         if len(self.output_file) == 1:
             self.output_file = self.output_file[0]
@@ -611,8 +484,6 @@ class Convert:
             Defaults to ``False``
         parallel : bool
             whether or not to use parallel processing. (Not yet implemented)
-        extra_platform_data : Dataset
-            The dataset containing the platform information to be added to the output
         storage_options : dict
             Additional keywords to pass to the filesystem class.
         """
@@ -639,8 +510,6 @@ class Convert:
             Defaults to ``False``
         parallel : bool
             whether or not to use parallel processing. (Not yet implemented)
-        extra_platform_data : Dataset
-            The dataset containing the platform information to be added to the output
         storage_options : dict
             Additional keywords to pass to the filesystem class.
         """
